requirements.txt
aiohttp >= 3.9.3
aiohttp-proxy >= 0.1.2
better-proxy >= 1.1.5
loguru >= 0.7.2
pycryptodomex >= 3.20.0
pydantic >= 2.8.2
pydantic-settings >= 2.2.1
Pyrogram >= 2.0.106
pytoniq >= 0.1.39
Requests >= 2.32.3
typing_extensions >= 4.12.2
